refactor : BaseEntity -> 추상클래스
