+++ src/main/java/kr/co/hhjpetclinicstudy/service/model/BaseEntity.java
@@ -1,4 +1,4 @@
-package kr.co.hhjpetclinicstudy.persistence;
+package kr.co.hhjpetclinicstudy.service.model;
 
 import jakarta.persistence.*;
 import lombok.EqualsAndHashCode;
@@ -7,7 +7,7 @@ import lombok.Getter;
 @MappedSuperclass
 @EqualsAndHashCode(of = "id", callSuper = false)
 @Getter
-public class BaseEntity {
+public abstract class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
